pyproject.toml
[project]
name = "itmo-x5-agent"
version = "0.1.0"
description = "X5 case solution"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepeval>=2.5.3",
    "gradio>=5.20.1",
    "gradio-client>=1.7.2",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "langfuse>=2.59.7",
    "llama-index>=0.12.23",
    "llama-index-callbacks-langfuse>=0.3.0",
    "llama-index-llms-openai-like>=0.3.4",
    "llama-index-utils-workflow>=0.3.0",
    "llama-index-vector-stores-qdrant>=0.4.3",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "pymorphy3>=2.0.3",
    "qdrant-client>=1.12.1",
    "seaborn>=0.13.2",
    "torch>=2.5.1",
    "vllm>=0.7.3",
]

[dependency-groups]
prod = [
    "langfuse>=2.59.7",
    "llama-index>=0.12.23",
    "llama-index-callbacks-langfuse>=0.3.0",
    "llama-index-llms-openai-like>=0.3.4",
    "llama-index-vector-stores-qdrant>=0.4.3",
    "qdrant-client>=1.12.1",
    "gradio>=5.20.1",
    "pymorphy3>=2.0.3",
    "gradio>=5.20.1",
    "pydantic-settings>=2.6.0",
]
